feat: add receive pages
