+++ components/screens/receive-addresses/SelectAddress.tsx
@@ -3,10 +3,19 @@ import useWalletManager from "@/hooks/useWalletManager";
 import { toEvmAddress } from "@/lib/utils";
 import AddressItem from "@/components/screens/receive-addresses/AddressItem";
 import kasIcon from "@/assets/images/kas-icon.svg";
-import kasplexIcon from "@/assets/images/kasplex-icon.png";
+import { useNavigate } from "react-router-dom";
+import { useSettings } from "@/hooks/useSettings";
+import { TESTNET_SUPPORTED_EVM_L2_CHAINS } from "@/lib/layer2";
+import { numberToHex } from "viem";
+import { getChainImage } from "@/lib/layer2";
 
 export default function SelectAddress() {
   const { account } = useWalletManager();
+  const navigate = useNavigate();
+  const [settings] = useSettings();
+
+  const supportEvmL2s =
+    settings?.networkId === "mainnet" ? [] : TESTNET_SUPPORTED_EVM_L2_CHAINS;
 
   const kasAddress = account?.address ?? "";
   const evmAddress = toEvmAddress(account?.publicKeys?.[0] ?? "");
@@ -19,12 +28,23 @@ export default function SelectAddress() {
           address={kasAddress}
           chainName="Kaspa"
           imageUrl={kasIcon}
+          redirect={() => navigate("/receive/kaspa")}
         />
-        <AddressItem
-          address={evmAddress}
-          chainName="Kasplex"
-          imageUrl={kasplexIcon}
-        />
+
+        {supportEvmL2s.map((chain) => {
+          const chainName = chain.name;
+          const chainIdHex = numberToHex(chain.id);
+          const chainIcon = getChainImage(chainIdHex);
+          return (
+            <AddressItem
+              key={chain.id}
+              address={evmAddress}
+              chainName={chainName}
+              imageUrl={chainIcon}
+              redirect={() => navigate(`/receive/evm/${chainIdHex}`)}
+            />
+          );
+        })}
       </div>
     </div>
   );
