+++ entrypoints/popup/router.tsx
@@ -11,7 +11,6 @@ import ConnectConfirm from "@/components/screens/browser-api/ConnectConfirm";
 import SignAndBroadcastTxConfirm from "@/components/screens/browser-api/kaspa/SignAndBroadcastTxConfirm";
 import Dashboard from "@/components/screens/Dashboard";
 import Settings from "@/components/screens/Settings.tsx";
-import Receive from "@/components/screens/Receive.tsx";
 import Onboarding from "@/components/screens/Onboarding.tsx";
 import AddWallet from "@/components/screens/AddWallet.tsx";
 import ImportRecoveryPhrase from "@/components/screens/full-pages/ImportRecoveryPhrase";
@@ -73,6 +72,8 @@ import Erc20Asset from "@/components/screens/Erc20Asset";
 import { EVMAssetsProvider } from "@/contexts/EvmAssets";
 import EvmKasAsset from "@/components/screens/EvmKasAsset";
 import SelectAddress from "@/components/screens/receive-addresses/SelectAddress";
+import KaspaReceiveAddress from "@/components/screens/receive-addresses/KaspaReceiveAddress";
+import EvmReceiveAddress from "@/components/screens/receive-addresses/EvmReceiveAddress";
 
 const loadKaspaWasm = async () => {
   await init({ module_or_path: kaspaModule });
@@ -212,7 +213,15 @@ export const router = createHashRouter([
                       </KRC721RecentTransferProvider>
                     ),
                   },
-                  { path: "receive", element: <SelectAddress /> },
+                  {
+                    path: "receive/select-address",
+                    element: <SelectAddress />,
+                  },
+                  { path: "receive/kaspa", element: <KaspaReceiveAddress /> },
+                  {
+                    path: "receive/evm/:chainId",
+                    element: <EvmReceiveAddress />,
+                  },
                   { path: "settings", element: <Settings /> },
                   {
                     path: "connected-apps",
